Fix test for tensor transform ops.
That op created an empty (and invalid) transform op with
`TransformOpInterface()` for testing the tracking listeners. The
previous listeners never accessed the op created this way, so this
wasn't a problem. However, the new version does, so we need to create a
real op, which is what this commit does.